Refactor All the Library
